backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
motor==3.3.2
pymongo==4.6.1
pydantic[email]==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
mangum==0.17.0
slowapi==0.1.9
faker==22.5.1
